backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.116.1",
    "mcp>=1.13.1",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "starlette>=0.47.2",
    "uvicorn>=0.35.0",
]
